requirements.txt
Django==2.0.12
django-debug-toolbar==1.9.1
django-filter==1.1.0
django_opentracing==1.0.1
djangorestframework==3.8.2
#gevent==1.3.3
#greenlet==0.4.13
gunicorn==19.8.1
jaeger-client==4.0.0
Markdown==2.6.11
psycopg2==2.7.7
pytz==2018.4
PyYAML==5.1
sqlparse==0.2.4
whitenoise==4.1.2
